requirements.txt
kafka-python
elasticsearch==8.13.0
